platform: msm: gsi: Export symbols only if compiled as module
This driver exports WAY TOO MANY symbols and there is no time to
sort this and that.
To fix dangerous/unsupported relocations in "modern" toolchains
(GCC newer than 4.7), just export all symbols only if the GSI is
compiled as a module (as this guarantees that all the users of
this driver will also be modules).

This is not right at all: it's still possible that we compile this
driver statically in the kernel and others as modules, which will
actually have GSI dependency, but whatever uses GSI is a core
driver in our embedded devices, so let's just not care during the
early porting stage.

Maybe, one day....